+++ src/Main.java
@@ -1,9 +1,13 @@
 
 public class Main {
 
+    private static final String ID = "본인 아이디";
+    private static final String PASSWORD = "본인 비밀번호";
+    private static final String INBOX = "INBOX";
+    private static final String UNSEEN = "UNSEEN";
+
     public static void main(String[] args) {
-        ShowNaverEmailList showNaverEmailList = new ShowNaverEmailList("본인아이디","비번",
-                "INBOX", "UNSEEN");
+        ShowNaverEmailList showNaverEmailList = new ShowNaverEmailList(ID,PASSWORD, INBOX, UNSEEN);
         showNaverEmailList.returnList();
 
         for (String s : showNaverEmailList.getEmailIds()){
